update wordcloud files
